SortedSets:zRangeByScore
