+++ src/Traits/SortedSets.php
@@ -315,9 +315,39 @@ trait SortedSets
         return $this->redis->zRangeByScore($key, $start, $end, $options);
     }
 
-    public function zRevRangeByScore(): bool
+    /**
+     * Returns the elements of the sorted set stored at the specified key which
+     * have scores in the range [start,end]. Adding a parenthesis before start
+     * or end excludes it from the range. +inf and -inf are also valid limits.
+     *
+     * zRevRangeByScore returns the same items in reverse order, when the start
+     * and end parameters are swapped.
+     *
+     * See: https://redis.io/commands/zrevrangebyscore.
+     *
+     * @param  string     $key
+     * @param  mixed|int|string     $start
+     * @param  mixed|int|string     $end
+     * @param  array|null           $options  Two options are available:
+     *                                        - withscores => TRUE,
+     *                                        and limit => [$offset, $count]
+     *
+     * @return array                        Array containing the values in
+     *                                      specified range.
+     */
+    public function zRevRangeByScore(string $key, $start, $end, ?array $options = null): array
     {
-        return false;
+        if (is_null($options)) {
+            return $this->redis->zRevRangeByScore($key, $start, $end);
+        }
+
+        $rangeOptions = ['withscores', 'limit'];
+
+        if (count(array_intersect(array_keys($options), $rangeOptions)) != count($options)) {
+            throw new UnsupportedOptionException("Option Not Supported", 1);
+        }
+
+        return $this->redis->zRevRangeByScore($key, $start, $end, $options);
     }
 
     public function zRangeByLex(): bool
